--- A3/test.py
@@ -1,9 +1,0 @@
-a = 'EEEESESSESEESSEESEEEEEEESEESESSSESEEESSSSSSSSESSSSEESSSSESESSEEESESSSSEESSSSSESSESSSSSEESESESEEESSSEESESEESSSSSSSSSEEEESEESSESESESSSESSSEEESEEEESSSEESSEEESEEESEEEEEESEEESSESSESESESSSEEESESEEEEEEEESS'
-b = 'EEEESESSESEESSEESEEEEEEESEESESSSESEEESSSSSSSSESSSSEESSSSESESSSSSSESEEEEESSSSSESSESSSSSEESESESEEESSSEESESEESSSSSSSSSEEEESEESSESESESSSESSSEEESEEEESSSEESSEEESEEESEEEEEESEEESSESSESESESSSEEESESEEEEEEEESS'
-
-if a == b:
-    print('OK')
-
-for idx, val in enumerate(a):
-    if val != b[idx]:
-        print(f'{val = }; {b[idx] = }; AT {idx}')
